# Optimize owners query