Get Safe by owner query is quite slow. A new query that takes half the time is proposed. Current query

```sql
SELECT DISTINCT(address)
                    FROM (
                        SELECT address, owners,
                                rank() OVER (PARTITION BY address ORDER BY nonce DESC, internal_tx_id DESC) AS pos
                        FROM history_safestatus
                        WHERE address IN (
                            SELECT address FROM history_safestatus
                            WHERE owners @> ARRAY[%s]::bytea[]
                        )
                        ) AS ss
                    WHERE pos = 1 AND owners @> ARRAY[%s]::bytea[];
```

```
Unique  (cost=13777.67..20363.74 rows=1 width=21) (actual time=47.230..55.125 rows=545 loops=1)
   ->  Subquery Scan on ss  (cost=13777.67..20363.73 rows=1 width=21) (actual time=47.229..54.961 rows=545 loops=1)
         Filter: ((ss.owners @> '{"ETHEREUM_ADDRESS"}'::bytea[]) AND (ss.pos = 1))
         Rows Removed by Filter: 255
         ->  WindowAgg  (cost=13777.67..19711.50 rows=43482 width=124) (actual time=47.223..54.645 rows=800 loops=1)
               ->  Gather Merge  (cost=13777.67..18841.86 rows=43482 width=116) (actual time=47.210..53.823 rows=800 loops=1)
                     Workers Planned: 2
                     Workers Launched: 2
                     ->  Sort  (cost=12777.64..12822.94 rows=18118 width=116) (actual time=38.443..38.472 rows=267 loops=3)
                           Sort Key: history_safestatus.address, history_safestatus.nonce DESC, history_safestatus.internal_tx_id DESC
                           Sort Method: quicksort  Memory: 50kB
                           Worker 0:  Sort Method: quicksort  Memory: 45kB
                           Worker 1:  Sort Method: quicksort  Memory: 101kB
                           ->  Hash Join  (cost=2510.42..11496.24 rows=18118 width=116) (actual time=5.955..38.078 rows=267 loops=3)
                                 Hash Cond: (history_safestatus.address = history_safestatus_1.address)
                                 ->  Parallel Seq Scan on history_safestatus  (cost=0.00..8612.88 rows=65288 width=116) (actual time=0.003..16.814 rows=54089 loops=3)
                                 ->  Hash  (cost=2501.34..2501.34 rows=726 width=21) (actual time=3.140..3.141 rows=545 loops=3)
                                       Buckets: 1024  Batches: 1  Memory Usage: 37kB
                                       ->  HashAggregate  (cost=2494.08..2501.34 rows=726 width=21) (actual time=2.943..3.014 rows=545 loops=3)
                                             Group Key: history_safestatus_1.address
                                             Batches: 1  Memory Usage: 105kB
                                             Worker 0:  Batches: 1  Memory Usage: 105kB
                                             Worker 1:  Batches: 1  Memory Usage: 105kB
                                             ->  Bitmap Heap Scan on history_safestatus history_safestatus_1  (cost=281.75..2492.23 rows=742 width=21) (actual time=2.015..2.675 rows=800 loops=3)
                                                   Recheck Cond: (owners @> '{"ETHEREUM_ADDRESS"}'::bytea[])
                                                   Heap Blocks: exact=490
                                                   ->  Bitmap Index Scan on history_saf_owners_295490_gin  (cost=0.00..281.56 rows=742 width=0) (actual time=1.956..1.956 rows=800 loops=3)
                                                         Index Cond: (owners @> '{"ETHEREUM_ADDRESS"}'::bytea[])
 Planning Time: 0.675 ms
 Execution Time: 55.245 ms
```

New proposed query:
```sql
  SELECT address
                    FROM (
                        SELECT DISTINCT ON(address) address, owners
                        FROM history_safestatus
                        WHERE address IN (
                            -- Do a first filtering
                            SELECT address FROM history_safestatus
                            WHERE owners @> ARRAY[%s]::bytea[]
                        )
                        ORDER BY address, nonce DESC, internal_tx_id DESC
                    ) AS sq
                    WHERE owners @> ARRAY[%s]::bytea[];
```

```
  Subquery Scan on sq  (cost=13937.67..19508.31 rows=75 width=21) (actual time=24.577..25.012 rows=545 loops=1)
   Filter: (sq.owners @> '{"ETHEREUM_ADDRESS"}'::bytea[])
   ->  Unique  (cost=13937.67..19321.38 rows=14955 width=108) (actual time=24.572..24.886 rows=545 loops=1)
         ->  Gather Merge  (cost=13937.67..19208.24 rows=45254 width=108) (actual time=24.571..24.792 rows=800 loops=1)
               Workers Planned: 2
               Workers Launched: 2
               ->  Sort  (cost=12937.64..12984.78 rows=18856 width=108) (actual time=20.012..20.036 rows=267 loops=3)
                     Sort Key: history_safestatus.address, history_safestatus.nonce DESC
                     Sort Method: quicksort  Memory: 57kB
                     Worker 0:  Sort Method: quicksort  Memory: 76kB
                     Worker 1:  Sort Method: quicksort  Memory: 62kB
                     ->  Hash Join  (cost=2575.05..11598.61 rows=18856 width=108) (actual time=3.056..19.706 rows=267 loops=3)
                           Hash Cond: (history_safestatus.address = history_safestatus_1.address)
                           ->  Parallel Seq Scan on history_safestatus  (cost=0.00..8636.27 rows=67627 width=108) (actual time=0.003..7.524 rows=54103 loops=3)
                           ->  Hash  (cost=2565.67..2565.67 rows=750 width=21) (actual time=2.392..2.394 rows=545 loops=3)
                                 Buckets: 1024  Batches: 1  Memory Usage: 37kB
                                 ->  HashAggregate  (cost=2558.17..2565.67 rows=750 width=21) (actual time=2.187..2.274 rows=545 loops=3)
                                       Group Key: history_safestatus_1.address
                                       Batches: 1  Memory Usage: 105kB
                                       Worker 0:  Batches: 1  Memory Usage: 105kB
                                       Worker 1:  Batches: 1  Memory Usage: 105kB
                                       ->  Bitmap Heap Scan on history_safestatus history_safestatus_1  (cost=281.95..2556.25 rows=768 width=21) (actual time=1.313..1.927 rows=800 loops=3)
                                             Recheck Cond: (owners @> '{"ETHEREUM_ADDRESS"}'::bytea[])
                                             Heap Blocks: exact=490
                                             ->  Bitmap Index Scan on history_saf_owners_295490_gin  (cost=0.00..281.76 rows=768 width=0) (actual time=1.256..1.257 rows=800 loops=3)
                                                   Index Cond: (owners @> '{"ETHEREUM_ADDRESS"}'::bytea[])
 Planning Time: 0.360 ms
 Execution Time: 25.081 ms
```

Other queries tested:

```sql
EXPLAIN ANALYZE SELECT address FROM (SELECT DISTINCT ON(address) address, owners FROM history_safestatus) AS sq WHERE owners @> ARRAY['ETHEREUM_ADDRESS']::bytea[];
```

```
 Subquery Scan on sq  (cost=32509.33..33507.82 rows=75 width=21) (actual time=164.321..247.866 rows=545 loops=1)
   Filter: (sq.owners @> '{"ETHEREUM_ADDRESS"}'::bytea[])
   Rows Removed by Filter: 32140
   ->  Unique  (cost=32509.33..33320.88 rows=14955 width=104) (actual time=164.253..237.489 rows=32685 loops=1)
         ->  Sort  (cost=32509.33..32915.11 rows=162311 width=104) (actual time=164.251..213.220 rows=162311 loops=1)
               Sort Key: history_safestatus.address
               Sort Method: external merge  Disk: 18232kB
               ->  Seq Scan on history_safestatus  (cost=0.00..9583.11 rows=162311 width=104) (actual time=0.006..42.572 rows=162311 loops=1)
 Planning Time: 0.136 ms
 Execution Time: 252.153 ms
```

```sql 
EXPLAIN ANALYZE SELECT address FROM history_safestatus WHERE owners @> ARRAY['ETHEREUM_ADDRESS']::bytea[] AND internal_tx_id IN (SELECT DISTINCT ON (address) internal_tx_id FROM history_safestatus ORDER BY address, nonce DESC, internal_tx_id DESC);
```  

```
 Nested Loop  (cost=10851.77..12445.53 rows=71 width=21) (actual time=53.150..137.612 rows=545 loops=1)
   ->  HashAggregate  (cost=10851.35..10853.35 rows=200 width=8) (actual time=52.923..59.590 rows=32688 loops=1)
         Group Key: history_safestatus_1.internal_tx_id
         Batches: 1  Memory Usage: 3105kB
         ->  Result  (cost=0.42..10664.41 rows=14955 width=33) (actual time=0.016..44.297 rows=32688 loops=1)
               ->  Unique  (cost=0.42..10664.41 rows=14955 width=33) (actual time=0.015..40.717 rows=32688 loops=1)
                     ->  Index Only Scan using test1 on history_safestatus history_safestatus_1  (cost=0.42..10258.58 rows=162333 width=33) (actual time=0.014..25.677 rows=162337 loops=1)
                           Heap Fetches: 22671
   ->  Index Scan using history_safestatus_pkey on history_safestatus  (cost=0.42..7.96 rows=1 width=29) (actual time=0.002..0.002 rows=0 loops=32688)
         Index Cond: (internal_tx_id = history_safestatus_1.internal_tx_id)
         Filter: (owners @> '{"ETHEREUM_ADDRESS"}'::bytea[])
         Rows Removed by Filter: 1
 Planning Time: 0.201 ms
 Execution Time: 137.680 ms
(14 rows)
```